test: refactor conftest.py into smaller parts
I'd like to share the vcs fixtures with other tests, but conftest.py is
getting a bit unwieldy and difficult to navigate. This creates a
'fixtures' directory where we can group like-minded fixtures together
into better organized modules.
